♻️ [Refactoring]: #140 [FE] selectTaskOutcome を src/lib/ 配下の専用モジュールへ切り出し App.tsx の公開境界を整理
Copilot レビュー指摘対応。`selectTaskOutcome` を `App.tsx` から export
していたが、テスト都合の内部ロジックが `@/App` の公開 API 面積を広げる
ため、`src/lib/selectTaskOutcome/index.ts` に切り出して App 側は import
で参照する形に変更する。CLAUDE.md の lib 配下フォルダ規約
（{Name}/index.ts + __tests__/）に整合させ、テストファイルも
`__tests__/selectTaskOutcome.behavior.test.ts` に移動する。

Refs #140
